fix(plugins): use Long.parseLong in isVersionCompatible to fix yyMMddHHmm overflow
isVersionCompatible() used Integer.parseInt() on version segments, but
DevRepoSAXHandler appends yyMMddHHmm timestamps (e.g. 2608090456) which
exceed Integer.MAX_VALUE (2147483647) for years >= 2022. This caused
NumberFormatException on every plugin refresh, crashing the
PluginRepoUpdater thread.

Switch to Long.parseLong() — same fix already applied to compareVersions
in 943d2224.

Validated on sagetv-mine container: plugin refresh now parses
1.0.0.2608091328 cleanly with no exceptions.

+++ java/sage/plugin/CorePluginManager.java
@@ -2937,8 +2937,8 @@ public class CorePluginManager implements Runnable
         int idx2b = minVer.indexOf('.');
         do
         {
-          int t = Integer.parseInt((idx1b == -1) ? target.substring(idx1a) : target.substring(idx1a, idx1b));
-          int m = Integer.parseInt((idx2b == -1) ? minVer.substring(idx2a) : minVer.substring(idx2a, idx2b));
+          long t = Long.parseLong((idx1b == -1) ? target.substring(idx1a) : target.substring(idx1a, idx1b));
+          long m = Long.parseLong((idx2b == -1) ? minVer.substring(idx2a) : minVer.substring(idx2a, idx2b));
           if (t < m)
             return false;
           if (t > m)
@@ -2962,8 +2962,8 @@ public class CorePluginManager implements Runnable
         int idx2b = maxVer.indexOf('.');
         do
         {
-          int t = Integer.parseInt((idx1b == -1) ? target.substring(idx1a) : target.substring(idx1a, idx1b));
-          int m = Integer.parseInt((idx2b == -1) ? maxVer.substring(idx2a) : maxVer.substring(idx2a, idx2b));
+          long t = Long.parseLong((idx1b == -1) ? target.substring(idx1a) : target.substring(idx1a, idx1b));
+          long m = Long.parseLong((idx2b == -1) ? maxVer.substring(idx2a) : maxVer.substring(idx2a, idx2b));
           if (t > m)
             return false;
           if (t < m)
